requirements.txt
Django==4.2
djangorestframework==3.14.0
flake8==6.1.0
black==23.10.1
flake8-black==0.3.6 
flake8-isort==6.0.0
pytest==7.4.0
pytest-django==4.7.0
pytest-cov==4.1.0  
factory-boy==3.3.0 
isort>=5.12.0